// src/api/generated/schemas/ticketLogResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente. Autenticação via cookie HttpOnly `worklog_access` emitido pelo endpoint `POST /worklog/auth/login` (o browser anexa o cookie automaticamente em chamadas subsequentes).
 * OpenAPI spec version: v1.0.0
 */
import type { TicketLogResponseFieldType } from './ticketLogResponseFieldType';
import type { UserSummary } from './userSummary';

export interface TicketLogResponse {
  changeGroupId?: string;
  fieldChanged?: string;
  fieldType?: TicketLogResponseFieldType;
  oldValue?: string;
  newValue?: string;
  changeDate?: string;
  user?: UserSummary;
}